Feature / Notifications (#18)
* feat: add notifications module with controller and service (copilot)

* feat: implement notification service and gateway with error handling

* feat: integrate notification service into AuthService for account creation notifications

* feat: add user status to user entity

* feat: delete notifications module

* feat: listen user creation

* feat: remove access token in register or when user is not active

* feat: notification to user with same level

* feat: remove notification enums

* refac: move notifications to the correct folder

* refac: remove circular dependency

* feat: send notification to all users

* feat: basic auth layer websocket

* style: format

* feat: create user automatically when is admin

* feat: convert entity to dto

* refac: from interface to class

* fix: missing .data when acessing ws user

* refac: delete unused imports

* feat: protect routes by role

* fix: tests

* feat: user can only update youserlf

* style: remove unused imports

* feat: all permissions set

---------

+++ src/modules/users/services/users.service.ts
@@ -1,10 +1,20 @@
-import { Injectable, NotFoundException, ConflictException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ConflictException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from '@users/entities/user.entity';
 import { UpdateUserDto } from '@users/dto/update-user.dto';
-import { RegisterUserDto } from '@users/dto/register-user.dto';
+import { CreateUserDto } from '@users/dto/create-user.dto';
 import { ILike } from 'typeorm';
+import UserLevel from '@users/entities/user-level';
+import UserStatus from '@users/entities/user-status';
+import * as bcrypt from 'bcrypt';
+import { UserDto } from '@users/dto/user.dto';
+import { JwtPayload } from '@users/dto/jwt-payload.interface';
 
 @Injectable()
 export class UsersService {
@@ -13,23 +23,43 @@ export class UsersService {
     private readonly userRepository: Repository<User>
   ) {}
 
-  async create(user: RegisterUserDto): Promise<User> {
+  async createAlreadyActive(user: CreateUserDto): Promise<UserDto> {
+    const newUser = Object.assign(new User(), {
+      ...user,
+      status: UserStatus.ACTIVE,
+    });
+
+    return this.create(newUser);
+  }
+
+  async create(user: CreateUserDto): Promise<UserDto> {
     await this.validateEmail(user.email, '');
     await this.validateUsername(user.username, '');
     await this.validateName(user.name, '');
-    return this.userRepository.save(user);
+
+    const encryptedPassword = await bcrypt.hash(user.password, 10);
+
+    user.password = encryptedPassword;
+
+    return this.userRepository.save(user).then(user => user.toDto());
   }
 
-  async findAll(): Promise<User[]> {
-    return this.userRepository.find();
+  async findAll(): Promise<UserDto[]> {
+    return (await this.userRepository.find()).map(user => user.toDto());
   }
 
-  async findOne(id: number): Promise<User> {
+  async findAllAdmins(): Promise<UserDto[]> {
+    return (await this.userRepository.find({ where: { level: UserLevel.ADMIN } })).map(user =>
+      user.toDto()
+    );
+  }
+
+  async findOne(id: number): Promise<UserDto> {
     const user = await this.userRepository.findOne({ where: { id } });
     if (!user) {
       throw new NotFoundException(`User with ID ${id} not found`);
     }
-    return user;
+    return user.toDto();
   }
 
   async findByLogin(login: string): Promise<User | undefined> {
@@ -38,18 +68,59 @@ export class UsersService {
     });
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
-    const user = await this.findOne(id);
+  async activeUser(id: number): Promise<UserDto> {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
+    user.activate();
+
+    return this.userRepository.save(user).then(user => user.toDto());
+  }
+
+  async update(
+    id: number,
+    updateUserDto: UpdateUserDto,
+    currentUser: JwtPayload
+  ): Promise<UserDto> {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
+    await this.validateUpdatePermission(currentUser, user, updateUserDto);
     await this.validateEmail(updateUserDto.email, user.email);
     await this.validateUsername(updateUserDto.username, user.username);
     await this.validateName(updateUserDto.name, user.name);
-    Object.assign(user, updateUserDto);
-    return this.userRepository.save(user);
+
+    return this.userRepository.update({ id }, updateUserDto).then(() => this.findOne(id));
   }
 
   async remove(id: number): Promise<void> {
-    const user = await this.findOne(id);
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+
     await this.userRepository.remove(user);
+  }
+
+  private async validateUpdatePermission(
+    currentUser: JwtPayload,
+    user: User,
+    updateUserDto: UpdateUserDto
+  ): Promise<void> {
+    if (currentUser.level !== UserLevel.ADMIN && currentUser.username !== user.username) {
+      throw new ForbiddenException('You can only update yourself');
+    }
+
+    if (currentUser.username === user.username && updateUserDto.level !== user.level) {
+      throw new ForbiddenException('You cannot update your own level');
+    }
   }
 
   private async validateEmail(newEmail: string | undefined, currentEmail: string): Promise<void> {
